requirements.txt
sqlalchemy
pydantic
python-dotenv
fastapi
uvicorn
art
